ezviz/pyproject.toml
[project]
name = "ezviz-devices"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "python-dotenv",
    "requests",
    "uvicorn",
]
